Interface and a semi-functioning asset manager

+++ src/IGameElement.hpp
@@ -6,7 +6,7 @@ class SDL_Renderer;
 class IGameElement
 {
 	public:
-		virtual void update(float delta_time) = 0;
+		virtual void update(int delta_time) = 0;
 		virtual void draw(SDL_Renderer *renderer) = 0;
 };
